feat: BigDecimal

+++ api/src/main/java/com/bindglam/goldengine/account/Account.java
@@ -2,14 +2,15 @@ package com.bindglam.goldengine.account;
 
 import org.jetbrains.annotations.NotNull;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface Account extends AutoCloseable {
     @NotNull UUID holder();
 
-    double balance();
+    BigDecimal balance();
 
-    void balance(double balance);
+    void balance(BigDecimal balance);
 
-    boolean modifyBalance(double amount, Operation operation);
+    boolean modifyBalance(BigDecimal amount, Operation operation);
 }
